Traffic

+++ control_plane/entropyanalyzer.py
@@ -8,7 +8,7 @@ class EntropyDNSAnalyzer(DNSAnalyzer):
     DNSAnalyzer child class that analyzes DNS query qnames for different levels of entropy (randomness)
     """
 
-    def analyze(self, dns_event_query: RecordEvent) -> int:
+    def analyze(self, dns_event_query: RecordEvent) -> float:
         return 2 
 
     def report(self) -> str:
